fix: doctests
