+++ crates/proof-of-sql-parser/src/posql_time/timezone.rs
@@ -71,3 +71,63 @@ impl fmt::Display for PoSQLTimeZone {
         }
     }
 }
+
+#[cfg(test)]
+mod timezone_parsing_tests {
+    use crate::posql_time::timezone;
+
+    #[test]
+    fn test_display_fixed_offset_positive() {
+        let timezone = timezone::PoSQLTimeZone::FixedOffset(4500); // +01:15
+        assert_eq!(format!("{}", timezone), "+01:15");
+    }
+
+    #[test]
+    fn test_display_fixed_offset_negative() {
+        let timezone = timezone::PoSQLTimeZone::FixedOffset(-3780); // -01:03
+        assert_eq!(format!("{}", timezone), "-01:03");
+    }
+
+    #[test]
+    fn test_display_utc() {
+        let timezone = timezone::PoSQLTimeZone::Utc;
+        assert_eq!(format!("{}", timezone), "00:00");
+    }
+}
+
+#[cfg(test)]
+mod timezone_offset_tests {
+    use crate::posql_time::{timestamp::PoSQLTimestamp, timezone};
+
+    #[test]
+    fn test_utc_timezone() {
+        let input = "2023-06-26T12:34:56Z";
+        let expected_timezone = timezone::PoSQLTimeZone::Utc;
+        let result = PoSQLTimestamp::try_from(input).unwrap();
+        assert_eq!(result.timezone, expected_timezone);
+    }
+
+    #[test]
+    fn test_positive_offset_timezone() {
+        let input = "2023-06-26T12:34:56+03:30";
+        let expected_timezone = timezone::PoSQLTimeZone::from_offset(12600); // 3 hours and 30 minutes in seconds
+        let result = PoSQLTimestamp::try_from(input).unwrap();
+        assert_eq!(result.timezone, expected_timezone);
+    }
+
+    #[test]
+    fn test_negative_offset_timezone() {
+        let input = "2023-06-26T12:34:56-04:00";
+        let expected_timezone = timezone::PoSQLTimeZone::from_offset(-14400); // -4 hours in seconds
+        let result = PoSQLTimestamp::try_from(input).unwrap();
+        assert_eq!(result.timezone, expected_timezone);
+    }
+
+    #[test]
+    fn test_zero_offset_timezone() {
+        let input = "2023-06-26T12:34:56+00:00";
+        let expected_timezone = timezone::PoSQLTimeZone::Utc; // Zero offset defaults to UTC
+        let result = PoSQLTimestamp::try_from(input).unwrap();
+        assert_eq!(result.timezone, expected_timezone);
+    }
+}
